+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 
 ### 1.12.1
 
+- Refactors Artwork view to use FlatList - kieran
+
 ### 1.12.0
 
 - Upgrade palette@4.13.3; styled-components@4.x